Move print function to debug module
